+++ frontend/app/(views)/(website)/opengraph-image.js
@@ -1,8 +1,6 @@
 /* * */
 
-import { allCardsData } from '@/components/review-2024/_data/cards';
-import OpenGraphViagem2024Default from '@/opengraph/OpenGraphViagem2024Default';
-import OpenGraphViagem2024Dynamic from '@/opengraph/OpenGraphViagem2024Dynamic';
+import OpenGraphDefault from '@/opengraph/OpenGraphDefault';
 import fs from 'fs';
 import { ImageResponse } from 'next/og';
 
@@ -14,7 +12,7 @@ export const contentType = 'image/png';
 
 /* * */
 
-export default async function Image(req) {
+export default async function Image() {
 	//
 
 	//
@@ -26,28 +24,10 @@ export default async function Image(req) {
 		{ data: fs.readFileSync(`${process.cwd()}/public/fonts/Inter-Bold.ttf`).buffer, name: 'Inter', style: 'normal', weight: 700 },
 	];
 
-	// B. Fetch data
-	const baseUrl = process.env.NEXT_PUBLIC_URL;
-	const url = new URL(req.url, baseUrl);
-	const cardId = url.searchParams.get('card') || 'unknown';
-
-	const cardData = allCardsData.find(item => item._id === cardId);
-
-	console.log('CAARD', url.pathname);
-
 	//
 	// C. Render default component
 
-	if (!cardData) {
-		return new ImageResponse(<OpenGraphViagem2024Default />, { ...size, fonts: customFonts });
-	}
-
-	//
-
-	//
-	// D. Render dynamic component
-
-	return new ImageResponse(<OpenGraphViagem2024Dynamic cardData={cardData} />, { ...size, fonts: customFonts });
+	return new ImageResponse(<OpenGraphDefault />, { ...size, fonts: customFonts });
 
 	//
 }
